currency_convertor
A universal currency converter is an app or web tool that allows for the quick
conversion of any currency into any other currency. Universal converters
typically use the most recent market prices in the foreign exchange market.

Currency converters aim to maintain real-time information on current market
or bank exchange rates, so that the calculated result changes whenever the
value of either of the component currencies does. They do so by connecting
to a database of current currency exchange rates.