word counter
The word count is the number of words in a document or passage of text.
Word counting may be needed when a text is required to stay within certain
numbers of words. This may particularly be the case in academia, legal
proceedings, journalism and advertising.
Word counts the number of words in a document while you type. Word also
counts pages, paragraphs, lines, and characters.