emailsystem
Electronic mail (e-mail) is a computer-based application for the
exchange of messages between users. A worldwide e-mail network
allows people to exchange e-mail messages very quickly. E-mail is the
electronic equivalent of a letter, but with advantages in timeliness and
flexibility.
mail is important for communication because it allows users to send
information in letter format, and email can replace traditional mail
options. Emails can be more beneficial for communication because they
can often include text, documents and multimedia, like photos and
videos.